fix(func): object_construct only ignore nil
Refactor to move the function
Do not return nil if encountering nil
